% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-PropertySales20012017.R
\docType{data}
\name{PropertySales20012017}
\alias{PropertySales20012017}
\title{Property Sales 2001-20017}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 3023750 rows and 24 columns.}
\usage{
PropertySales20012017
}
\description{
All property sales recorded by the NSW Valuer-General
}
\keyword{datasets}
